pyproject.toml
[project]
name = "linkedin-scraper"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "playwright>=1.51.0",
    "pydantic>=2.11.3",
    "python-dotenv>=1.1.0",
    "tiktoken>=0.9.0",
    "together>=1.5.5",
    "undetected-playwright>=0.3.0",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
    "pytest>=8.3.5",
]
